sala's controller and service created and tested

+++ src/salas/salas.service.ts
@@ -1,5 +1,5 @@
 /* eslint-disable prettier/prettier */
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { Prisma } from '@prisma/client';
 import { PrismaService } from 'src/prisma.service';
 
@@ -7,28 +7,63 @@ import { PrismaService } from 'src/prisma.service';
 export class SalasService {
   constructor(private prisma: PrismaService){}
 
-  create(createSalaDto: Prisma.SalaUncheckedCreateInput) {
+  async create(createSalaDto: Prisma.SalaUncheckedCreateInput) {
+    
+    const numero = createSalaDto.numero;
+    const sala = await this.prisma.sala.findUnique({
+      where: {numero}
+    })
+
+    if(sala){
+      throw new HttpException('Já existe uma sala com esse número', HttpStatus.BAD_REQUEST);
+    }
+
+    
     return this.prisma.sala.create({
       data: createSalaDto
     })
   }
 
-  findOne(numero: number) {
-    return this.prisma.sala.findUnique({
-      where:{ numero }
+  async findOne(where: Prisma.SalaWhereUniqueInput) {
+    const sala = await this.prisma.sala.findUnique({
+      where
     })
+
+    if(!sala){
+      throw new HttpException('Sala não encontrada', HttpStatus.BAD_REQUEST);
+    } else {
+      return sala;
+    }
   }
 
-  update(numero: number, updateSalaDto: Prisma.SalaUpdateInput) {
+  async update(where: Prisma.SalaWhereUniqueInput, data: Prisma.SalaUpdateInput) {
+    const sala = await this.prisma.sala.findUnique({
+      where
+    })
+
+    if(!sala){
+      throw new HttpException('Sala não encontrada', HttpStatus.BAD_REQUEST)
+    }
+
     return this.prisma.sala.update({
-      where:{numero},
-      data: updateSalaDto
+      where,
+      data
     })
+
+
   }
 
-  remove(numero: number) {
+  async remove(where: Prisma.SalaWhereUniqueInput) {
+
+    const sala = await this.prisma.sala.findUnique({
+      where
+    })
+
+    if(!sala){
+      throw new HttpException('Sala não encontrada', HttpStatus.BAD_REQUEST)
+    }
     return this.prisma.sala.delete({
-      where: {numero}
+      where
     })
   }
 }
